requirements.txt
asgiref==3.6.0
autopep8==2.0.1
axios==0.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dj-database-url==2.0.0
Django==4.1.4
django-cors-headers==4.0.0
django-environ==0.10.0
djangorestframework==3.14.0
docopt==0.6.2
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
lxml==4.9.2
markdown-it-py==2.2.0
mccabe==0.7.0
mdurl==0.1.2
mysql-connector-python==8.0.32
node==1.2
odict==1.9.0
Pillow==9.3.0
pipreqs==0.4.11
plumber==1.7
protobuf==3.20.3
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
pytz==2022.7.1
requests==2.28.2
rich==13.3.1
source==1.2.0
sqlparse==0.4.3
svgwrite==1.4.3
sweetify==2.3.1
Tree==0.2.4
typing_extensions==4.6.3
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.4.0
yarg==0.1.9
zope.component==5.1.0
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.6
zope.hookable==5.4
zope.interface==5.5.2
zope.lifecycleevent==4.4
zope.proxy==5.0.0
